Added some list code & reorganized some stuff
Wrote the init, add and get tail funciton of the linked list for our
command stream thing

also cleaned up some of my header files
