+++ stack.h
@@ -25,55 +25,21 @@ typedef struct{
 
 // STACK OPERATIONS:
 //create stack
-void newStack(stack *st, int dataSize){
-	assert(dataSize > 0);
-	st->dataSize = dataSize;
-	st->length = 0;
-	st->allocLength = initialSize;
-	st->data = malloc(initialSize * dataSize);
-	assert(st->data != NULL);
-} 
+void newStack(stack *st, int dataSize);
 
 //free stack
-void destroyStack(stack *st){
-	free(st->data);
-} 
+void destroyStack(stack *st);
 
 //isempty
-bool StackisEmpty(const stack *st){
-	if(st->length == 0) return true;
-	return false;
-} 
+bool StackisEmpty(const stack *st);
 
  //push
-void pushStack(stack *st, const void* add){
-	void *destination;
-	if(st->length == st->allocLength){
-		st->allocLength *= 2;
-		st->data = realloc(st->data, st->allocLength * st->dataSize);
-		assert(st->data != NULL);
-	}
-
-	destination = (char *)st->data + st->length * st->dataSize;
-	memcpy(destination,add,st->dataSize);
-	st->length++;
-}
+void pushStack(stack *st, const void* add);
 
 //pop
-void popStack(stack *st, void* top){
-	const void *src;
-	assert(!StackisEmpty(st));
-	st->length--;
-	src = (const char *) st->data + st->length * st->dataSize;
-	memcpy(top, src, st->dataSize);
-} 
+void popStack(stack *st, void* top);
 
 //view top element (doesnt affect stack)
-void topStack(stack *st, void* top){
-	const void *src;
-	assert(!StackisEmpty(st));
-	src = (const char *) st->data + (st->length - 1) * st->dataSize;
-	memcpy(top, src, st->dataSize);
-}
+void topStack(stack *st, void* top);
 
 #endif
